Remove unnecessary converson in ExecuteA/A64.hs
Variable x has "type" int32 or int64.
Variable y has "type" Register.
We want to store something of "type" int32 or int64.
So the only necessary conversion is the one converting y to int32/int64.

After calculating the value we don’t have to convert int32/int64 to
int32/int64 as if it had been a value of "type" Register.